refactor: 수강생 목록 헤더, 컨텐츠 컴포넌트 분리
